+++ app/com/lynxanalytics/biggraph/frontend_operations/WorkflowOperations.scala
@@ -609,7 +609,7 @@ class WorkflowOperations(env: SparkFreeEnvironment) extends ProjectOperations(en
         input.kind match {
           case BoxOutputKind.Project =>
             val project = projectInput("graph")
-            if (Seq("html", "ipywidget", "matplotlib").contains(params("outputs"))) {
+            if (Seq("html", "matplotlib").contains(params("outputs"))) {
               val html = PythonUtilities.deriveHTML(params("code"), params("outputs"), pythonInputs, project)
               // The output is called "graph" to preserve compatibility.
               Map(context.box.output("graph") -> BoxOutputState.html(html))
@@ -621,18 +621,24 @@ class WorkflowOperations(env: SparkFreeEnvironment) extends ProjectOperations(en
             // We named the input and output before adding table support.
             // It's bad naming here, but lets us keep compatibility.
             val table = tableInput("graph")
-            val outputs: Seq[String] =
-              if (params("outputs") == "<infer from code>")
-                table.schema.fields.map { f =>
-                  s"df.${f.name}: " + (f.dataType match {
-                    case org.apache.spark.sql.types.StringType => "str"
-                    case org.apache.spark.sql.types.LongType => "int"
-                    case _ => "float"
-                  })
-                } ++ pythonOutputs
-              else pythonOutputs
-            val result = PythonUtilities.deriveTable(params("code"), table, outputs)
-            Map(context.box.output("graph") -> BoxOutputState.from(result))
+            if (Seq("html", "matplotlib").contains(params("outputs"))) {
+              val html = PythonUtilities.deriveHTML(params("code"), params("outputs"), table)
+              // The output is called "graph" to preserve compatibility.
+              Map(context.box.output("graph") -> BoxOutputState.html(html))
+            } else {
+              val outputs: Seq[String] =
+                if (params("outputs") == "<infer from code>")
+                  table.schema.fields.map { f =>
+                    s"df.${f.name}: " + (f.dataType match {
+                      case org.apache.spark.sql.types.StringType => "str"
+                      case org.apache.spark.sql.types.LongType => "int"
+                      case _ => "float"
+                    })
+                  } ++ pythonOutputs
+                else pythonOutputs
+              val result = PythonUtilities.deriveTable(params("code"), table, outputs)
+              Map(context.box.output("graph") -> BoxOutputState.from(result))
+            }
         }
       }
       // Unused because we are overriding getOutputs.
@@ -649,8 +655,8 @@ class WorkflowOperations(env: SparkFreeEnvironment) extends ProjectOperations(en
     List("output"),
     new SmartOperation(_) {
       params ++= List(
-        Param("inputs", "Inputs", defaultValue = "<infer from code>"),
-        Param("outputs", "Outputs", defaultValue = "<infer from code>"),
+        Param("inputs", "Inputs"),
+        Param("outputs", "Outputs"),
         Code("code", "R code", language = "r"),
       )
       val input = context.inputs("input")
@@ -677,17 +683,22 @@ class WorkflowOperations(env: SparkFreeEnvironment) extends ProjectOperations(en
             }
           case BoxOutputKind.Table =>
             val table = tableInput("input")
-            val outputs: Seq[String] =
-              table.schema.fields.map { f =>
-                s"df.${f.name}: " + (f.dataType match {
-                  case org.apache.spark.sql.types.StringType => "character"
-                  case org.apache.spark.sql.types.LongType => "integer"
-                  case _ => "double"
-                })
-              } ++ rOutputs
-            log.error(s"DeriveR outputs: $outputs")
-            val result = graph_operations.DeriveR.deriveTable(params("code"), table, outputs)
-            Map(context.box.output("output") -> BoxOutputState.from(result))
+            if (Seq("html", "plot").contains(params("outputs"))) {
+              val html = graph_operations.DeriveR.deriveHTML(params("code"), params("outputs"), table)
+              Map(context.box.output("output") -> BoxOutputState.html(html))
+            } else {
+              val outputs: Seq[String] =
+                table.schema.fields.map { f =>
+                  s"df.${f.name}: " + (f.dataType match {
+                    case org.apache.spark.sql.types.StringType => "character"
+                    case org.apache.spark.sql.types.LongType => "integer"
+                    case _ => "double"
+                  })
+                } ++ rOutputs
+              log.error(s"DeriveR outputs: $outputs")
+              val result = graph_operations.DeriveR.deriveTable(params("code"), table, outputs)
+              Map(context.box.output("output") -> BoxOutputState.from(result))
+            }
         }
       }
       // Unused because we are overriding getOutputs.