+++ sphynx/python/derive_html.py
@@ -11,25 +11,29 @@ if os.environ.get('SPHYNX_CHROOT_PYTHON') == 'yes':
   op.run_in_chroot()
 
 # Load inputs.
-vs = {}
-es = {}
-graph_attributes = types.SimpleNamespace()
-for fullname in op.inputs.keys():
-  if '.' not in fullname:
-    continue
-  parent, name = fullname.split('.')
-  if parent == 'vs':
-    vs[name] = op.input(fullname)
-  elif parent == 'es':
-    es[name] = op.input(fullname)
-  elif parent == 'graph_attributes':
-    setattr(graph_attributes, name, op.input_scalar(fullname))
-if 'edges-for-es' in op.inputs:
-  edges = op.input('edges-for-es')
-  es['src'] = edges.src
-  es['dst'] = edges.dst
-vs = pd.DataFrame(vs)
-es = pd.DataFrame(es)
+if op.classname.endswith('DeriveHTMLPython'):
+  vs = {}
+  es = {}
+  graph_attributes = types.SimpleNamespace()
+  for fullname in op.inputs.keys():
+    if '.' not in fullname:
+      continue
+    parent, name = fullname.split('.')
+    if parent == 'vs':
+      vs[name] = op.input(fullname)
+    elif parent == 'es':
+      es[name] = op.input(fullname)
+    elif parent == 'graph_attributes':
+      setattr(graph_attributes, name, op.input_scalar(fullname))
+  if 'edges-for-es' in op.inputs:
+    edges = op.input('edges-for-es')
+    es['src'] = edges.src
+    es['dst'] = edges.dst
+  vs = pd.DataFrame(vs)
+  es = pd.DataFrame(es)
+elif op.classname.endswith('DeriveHTMLTablePython'):
+  for name in op.inputs:
+    globals()[name] = op.input_table(name)
 
 # Execute user code.
 try:
@@ -45,32 +49,11 @@ except BaseException:
 
 if op.params['mode'] == 'html':
   assert 'html' in globals(), 'Please save the output as "html".'
-  sc = html
 elif op.params['mode'] == 'matplotlib':
   import base64
   from matplotlib import pyplot as plt
   f = io.BytesIO()
   plt.savefig(f, format='svg')
   data = base64.b64encode(f.getvalue()).decode('ascii')
-  sc = f'<img src="data:image/svg+xml;base64, {data}">'
-elif op.params['mode'] == 'ipywidget':
-  from ipywidgets.embed import embed_minimal_html
-  f = io.StringIO()
-  embed_minimal_html(f, views=None)
-  sc = f.getvalue()
-  # The widget code has to be loaded from a CDN.
-  sc = sc.replace('embed-amd.js"', 'embed-amd.js" data-jupyter-widgets-cdn-only')
-  # We patch the URL class because this is displayed via <iframe srcdoc="...">.
-  sc = sc.replace('<body>', '''
-    <body>
-    <script>
-      const RURL = URL;
-    </script>
-    <script>
-      class URL extends RURL {
-        constructor(url, base) {
-          super(url, 'http://example.com/');
-        }
-      }
-    </script>''')
-op.output_scalar('sc', sc)
+  html = f'<img src="data:image/svg+xml;base64, {data}">'
+op.output_scalar('sc', html)