+++ sphynx/r/derive_html.r
@@ -1,41 +1,19 @@
 # Run user code on a graph and produce HTML output.
 source("r/util.r")
 
-ip = as.data.frame(installed.packages()[,c(1,3:4)])
-ip = ip[is.na(ip$Priority),1:2,drop=FALSE]
-print("packages")
-print(ip)
-
 # Create input tables.
-gettable <- function(parent) {
-    fields <- params[["inputFields"]]
-    special <- parent == "es" & (fields$name == "src" | fields$name == "dst")
-    ns <- fields[fields$parent == parent & !special, ]$name
-    if (length(ns) != 0) {
-        columns <- input_table(paste(parent, ns, sep = "."))
-        # .name_repair='minimal' allows duplicate column names. They are all
-        # called 'values'.
-        t <- tibble(!!!columns, .name_repair = "minimal")
-        # We set the correct column names here.
-        names(t) <- ns
-        t
+if (op[["Operation"]][["Class"]] == "com.lynxanalytics.biggraph.graph_operations.DeriveHTMLR") {
+    vs <- gettable("vs")
+    es <- gettable("es")
+    if ("edges-for-es" %in% names(inputs)) {
+        es <- cbind(es, input_edges("edges-for-es"))
+    }
+    graph_attributes <- getscalars()
+} else if (op[["Operation"]][["Class"]] == "com.lynxanalytics.biggraph.graph_operations.DeriveHTMLTableR") {
+    for (t in names(inputs)) {
+        assign(t, input_parquet(t))
     }
 }
-vs <- gettable("vs")
-es <- gettable("es")
-if ("edges-for-es" %in% names(inputs)) {
-    es <- cbind(es, input_edges("edges-for-es"))
-}
-
-getscalars <- function() {
-    fields <- params[["inputFields"]]
-    ns <- fields[fields$parent == "graph_attributes", ]$name
-    if (length(ns) != 0) {
-        values <- input_scalar(paste("graph_attributes", ns, sep = "."))
-        as.list(setNames(values, ns))
-    }
-}
-graph_attributes <- getscalars()
 
 if (params[["mode"]] == "plot") {
     svgfile <- tempfile()
@@ -45,13 +23,16 @@ if (params[["mode"]] == "plot") {
 # Run user code.
 print("RUNNING USER CODE")  # For log cleanup.
 code <- params[["code"]]
-eval(parse(text = code))
+html <- eval(parse(text = code))
 print("USER CODE FINISHED")
 
 # Save outputs.
 if (params[["mode"]] == "html") {
-    if (!exists("html")) {
-        stop("Please save the output as 'html'.")
+    if (!is.character(html)) {
+        library(htmlwidgets)
+        f <- tempfile()
+        htmlwidgets::saveWidget(html, f)
+        html <- readChar(f, file.info(f)$size)
     }
 } else if (params[["mode"]] == "plot") {
     library(openssl)